improve example code
